requirements.txt
fastapi
uvicorn
requests
jinja2
aiofiles
python-dotenv
python-multipart
twilio
